@@ -5,8 +5,7 @@ author: GwonHyeok
 homepage: https://github.com/GwonHyeok/flutter_channel_io
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
-  flutter: ">=1.20.0 <2.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
